The sun now goes "behind" the horizon
Two days worth of trying stuff. Turns out modifying a sprite during HBlank does nothing but make things randomly flicker. And using the VCount interrupt doesn't seem to do anything. So now this adapts the disintigation effect used for mirage tower to modify the sprite graphics in the sprite callback, so the sun sprite is literally erased line by line as the sun goes down, to simulate it going behind the bottommost background. It the GBA had a 4th background layer, this wouldn't even need to be done.